Add files via upload
These changes are for the new Differential Extruder. This extruder combines the advantages of the Bowden extruder (light print head) and Direct Drive extruder (precise extrusion). It works by adding the E motion to X motion and then sends the sum to the E stepper. The E stepper engages the extrusion wheel via a loop timing belt, so that only the difference between X and E motion rotates the extrusion wheel.
The changes in planner.cpp deal with printing, and changes in stepper.cpp deal with homing and manual commands issued from the printer console / screen.
Each change is preceded by the comment text "Beginning of a Differential Extruder block" and denoted by the sequence number. There are 6 changes/blocks in total.