Update planner.cpp
When the job timer is running, i.e. when a print job is active, planner.cpp adds X movement to and E movement and send the sum to the E stepper. 
When the job timer is not running, i.e. when commands can only come from the printer's console/screen, all processing specific to the differential extruder is done by stepper.cpp.